refactor: add request/response sub packagin in dto/note
to clarify each Dto's purpose
