--- src/main/java/com/example/teampandanback/dto/note/KanbanNoteSearchResponseDto.java
@@ -1,19 +1,0 @@
-package com.example.teampandanback.dto.note;
-
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-public class KanbanNoteSearchResponseDto {
-    private List<NoteOfProjectResponseDto> projects;
-
-    @Builder
-    public KanbanNoteSearchResponseDto(List<NoteOfProjectResponseDto> noteOfProjectResponseDtoList) {
-        this.projects = noteOfProjectResponseDtoList;
-    }
-
-}
-
