--- src/main/java/com/example/teampandanback/dto/note/NoteRequestDto.java
@@ -1,11 +1,0 @@
-package com.example.teampandanback.dto.note;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-public class NoteRequestDto {
-    private String title;
-    private String content;
-    private String deadline;
-}
